+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.2.3
+
+- Add `lockChannelsOrder` parameter to `ChannelsBloc`
+
+## 0.2.2+3
+
+- Fix `ChannelListView` channel hidden behaviour
+
+- Refresh `ChannelListView` on new message from hidden channel
+
+## 0.2.2+1
+
+- Fix some components to implement a splitview example
+
 ## 0.2.2
 
 - Add `messageLinks` property to `MessageTheme` to customize links color